Move IndividualTimePicker widget to its own file
